+++ matomo.php
@@ -4,7 +4,7 @@
  * Description: The #1 Google Analytics alternative that gives you full control over your data and protects the privacy for your users. Free, secure and open.
  * Author: Matomo
  * Author URI: https://matomo.org
- * Version: 4.3.1
+ * Version: 4.4.1
  * Domain Path: /languages
  * WC requires at least: 2.4.0
  * WC tested up to: 5.5.0
@@ -179,7 +179,7 @@ function matomo_add_plugin( $plugins_directory, $wp_plugin_file, $is_marketplace
 		}
 	}
 
-	$matomo_dir       = __DIR__ . '/app';
+	$matomo_dir       = __DIR__ . DIRECTORY_SEPARATOR . 'app';
 	$matomo_dir_parts = explode( DIRECTORY_SEPARATOR, $matomo_dir );
 	$root_dir_parts   = explode( DIRECTORY_SEPARATOR, $root_dir );
 	$webroot_dir      = '';