+++ tests/phpunit/wpmatomo/trackingcode/test-trackingcodegenerator.php
@@ -5,6 +5,7 @@
 
 use WpMatomo\Access;
 use WpMatomo\Admin\TrackingSettings;
+use WpMatomo\Admin\CookieConsent;
 use WpMatomo\Capabilities;
 use WpMatomo\Roles;
 use WpMatomo\Settings;
@@ -49,10 +50,11 @@ class TrackingCodeGeneratorTest extends MatomoUnit_TestCase {
 				'track_mode' => TrackingSettings::TRACK_MODE_DEFAULT,
 			)
 		);
+
 		$this->assertSame(
-			'<!-- Matomo --><script  >var _paq = window._paq = window._paq || [];
+			'<!-- Matomo --><script '.$this->get_type_attribute().'>'."\n".'var _paq = window._paq = window._paq || [];
 _paq.push([\'trackPageView\']);_paq.push([\'enableLinkTracking\']);_paq.push([\'alwaysUseSendBeacon\']);_paq.push([\'setTrackerUrl\', "\/\/example.org\/wp-content\/plugins\/matomo\/app\/matomo.php"]);_paq.push([\'setSiteId\', \'21\']);var d=document, g=d.createElement(\'script\'), s=d.getElementsByTagName(\'script\')[0];
-g.type=\'text/javascript\'; g.async=true; g.src="\/\/example.org\/wp-content\/plugins\/matomo\/app\/matomo.js"; s.parentNode.insertBefore(g,s);</script><!-- End Matomo Code -->',
+g.type=\'text/javascript\'; g.async=true; g.src="\/\/example.org\/wp-content\/plugins\/matomo\/app\/matomo.js"; s.parentNode.insertBefore(g,s);'."\n</script>\n<!-- End Matomo Code -->",
 			$this->get_tracking_code()
 		);
 	}
@@ -74,14 +76,14 @@ g.type=\'text/javascript\'; g.async=true; g.src="\/\/example.org\/wp-content\/pl
 		);
 
 		$this->assertSame(
-			'<!-- Matomo --><script  >var _paq = window._paq = window._paq || [];
+			'<!-- Matomo --><script '.$this->get_type_attribute().'>'."\n".'var _paq = window._paq = window._paq || [];
 _paq.push([\'addDownloadExtensions\', "zip|waf"]);
 _paq.push([\'setLinkClasses\', "clickme|foo"]);
 _paq.push([\'disableCookies\']);
 _paq.push([\'enableCrossDomainLinking\']);
 _paq.push(["setCookieDomain", "*.example.org"]);
 _paq.push([\'trackAllContentImpressions\']);_paq.push([\'trackPageView\']);_paq.push([\'enableLinkTracking\']);_paq.push([\'alwaysUseSendBeacon\']);_paq.push([\'setTrackerUrl\', "\/\/example.org\/index.php?rest_route=\/matomo\/v1\/hit\/"]);_paq.push([\'setSiteId\', \'21\']);var d=document, g=d.createElement(\'script\'), s=d.getElementsByTagName(\'script\')[0];
-g.type=\'text/javascript\'; g.async=true; g.src="\/\/example.org\/index.php?rest_route=\/matomo\/v1\/hit\/"; s.parentNode.insertBefore(g,s);</script><!-- End Matomo Code -->',
+g.type=\'text/javascript\'; g.async=true; g.src="\/\/example.org\/index.php?rest_route=\/matomo\/v1\/hit\/"; s.parentNode.insertBefore(g,s);'."\n</script>\n<!-- End Matomo Code -->",
 			$this->get_tracking_code()
 		);
 	}
@@ -210,5 +212,56 @@ g.type=\'text/javascript\'; g.async=true; g.src="http://example.org/wp-content/u
 		}
 	}
 
+	public function test_cookie_consent_tagmanager() {
+		$this->settings->apply_tracking_related_changes(
+			array(
+				'track_mode'     => TrackingSettings::TRACK_MODE_TAGMANAGER,
+				'cookie_consent' => CookieConsent::REQUIRE_COOKIE_CONSENT,
+			)
+		);
+		$this->assertNotContains( "requireCookieConsent", $this->get_tracking_code() );
+		$this->assertNotContains( "requireConsent", $this->get_tracking_code() );
+	}
 
+	public function test_cookie_consent_manually() {
+		$this->settings->apply_tracking_related_changes(
+			array(
+				'track_mode'     => TrackingSettings::TRACK_MODE_MANUALLY,
+				'cookie_consent' => CookieConsent::REQUIRE_COOKIE_CONSENT,
+			)
+		);
+		$this->assertNotContains( "requireCookieConsent", $this->get_tracking_code() );
+		$this->assertNotContains( "requireConsent", $this->get_tracking_code() );
+	}
+
+	public function test_cookie_consent_none() {
+		$this->settings->apply_tracking_related_changes(
+			array(
+				'track_mode'     => TrackingSettings::TRACK_MODE_DEFAULT,
+				'cookie_consent' => CookieConsent::REQUIRE_NONE,
+			)
+		);
+		$this->assertNotContains( "requireCookieConsent", $this->get_tracking_code() );
+		$this->assertNotContains( "requireConsent", $this->get_tracking_code() );
+	}
+
+	public function test_cookie_consent_cookie() {
+		$this->settings->apply_tracking_related_changes(
+			array(
+				'track_mode'     => TrackingSettings::TRACK_MODE_DEFAULT,
+				'cookie_consent' => CookieConsent::REQUIRE_COOKIE_CONSENT,
+			)
+		);
+		$this->assertContains( "_paq.push(['requireCookieConsent']);", $this->get_tracking_code() );
+	}
+
+	public function test_cookie_consent_tracking() {
+		$this->settings->apply_tracking_related_changes(
+			array(
+				'track_mode'     => TrackingSettings::TRACK_MODE_DEFAULT,
+				'cookie_consent' => CookieConsent::REQUIRE_TRACKING_CONSENT,
+			)
+		);
+		$this->assertContains( "_paq.push(['requireConsent']);", $this->get_tracking_code() );
+	}
 }